Directory testing contains a little test environment for an MMP It actually should provide a decent model for what we need in an actual MMP Send obviously can't just call receive, it'll have to package stuff up and send it, but if the receiver can handle a socket reception it should look a lot alike.
Host has myDataSets, which is a dictionary of (gasp!) data sets. Data sets have names
and a dictionary with index inputs and float outputs (floats cuz chicken, I guess).
Hosts can send given the name of the dataset, the index to send, and a host to send to.

Chicken chicken!